Major refactor: * Use uv for build * Separate into submodules * Add first test
