--- pylsl/__init__.py
@@ -1,51 +1,0 @@
-# (not using import * for Python 2.5 support)
-from .pylsl import (
-    DEDUCED_TIMESTAMP,
-    FOREVER,
-    IRREGULAR_RATE,
-    ContinuousResolver,
-    InternalError,
-    InvalidArgumentError,
-    LostError,
-    StreamInfo,
-    StreamInlet,
-    StreamOutlet,
-    TimeoutError,
-    XMLElement,
-    cf_double64,
-    cf_float32,
-    cf_int8,
-    cf_int16,
-    cf_int32,
-    cf_int64,
-    cf_string,
-    cf_undefined,
-    library_info,
-    library_version,
-    local_clock,
-    lost_error,
-    proc_ALL,
-    proc_clocksync,
-    proc_dejitter,
-    proc_monotonize,
-    proc_none,
-    proc_threadsafe,
-    protocol_version,
-    resolve_bypred,
-    resolve_byprop,
-    resolve_stream,
-    resolve_streams,
-    stream_info,
-    stream_inlet,
-    stream_outlet,
-    timeout_error,
-    vectorc,
-    vectord,
-    vectorf,
-    vectori,
-    vectorl,
-    vectors,
-    vectorstr,
-    xml_element,
-)
-from .version import __version__
