--- setup.cfg
@@ -1,8 +1,0 @@
-[metadata]
-# This includes the license file in the wheel.
-license_files = LICENSE
-
-[bdist_wheel]
-# This flag says that the code is written to work on both Python 2 and Python
-# 3.
-universal=1
