+++ src/pylsl/examples/PerformanceTest.py
