+++ src/pylsl/examples/README.md
