+++ src/pylsl/examples/ReceiveData.py
@@ -1,13 +1,12 @@
-"""Example program to demonstrate how to read a multi-channel time-series
-from LSL in a chunk-by-chunk manner (which is more efficient)."""
+"""Example program to show how to read a multi-channel time series from LSL."""
 
-from pylsl import StreamInlet, resolve_stream
+from pylsl import StreamInlet, resolve_byprop
 
 
 def main():
     # first resolve an EEG stream on the lab network
     print("looking for an EEG stream...")
-    streams = resolve_stream("type", "EEG")
+    streams = resolve_byprop("type", "EEG")
 
     # create a new inlet to read from the stream
     inlet = StreamInlet(streams[0])
@@ -15,9 +14,8 @@ def main():
     while True:
         # get a new sample (you can also omit the timestamp part if you're not
         # interested in it)
-        chunk, timestamps = inlet.pull_chunk()
-        if timestamps:
-            print(timestamps, chunk)
+        sample, timestamp = inlet.pull_sample()
+        print(timestamp, sample)
 
 
 if __name__ == "__main__":
