+++ src/pylsl/examples/ReceiveDataInChunks.py
@@ -1,12 +1,13 @@
-"""Example program to show how to read a multi-channel time series from LSL."""
+"""Example program to demonstrate how to read a multi-channel time-series
+from LSL in a chunk-by-chunk manner (which is more efficient)."""
 
-from pylsl import StreamInlet, resolve_stream
+from pylsl import StreamInlet, resolve_byprop
 
 
 def main():
     # first resolve an EEG stream on the lab network
     print("looking for an EEG stream...")
-    streams = resolve_stream("type", "EEG")
+    streams = resolve_byprop("type", "EEG")
 
     # create a new inlet to read from the stream
     inlet = StreamInlet(streams[0])
@@ -14,8 +15,9 @@ def main():
     while True:
         # get a new sample (you can also omit the timestamp part if you're not
         # interested in it)
-        sample, timestamp = inlet.pull_sample()
-        print(timestamp, sample)
+        chunk, timestamps = inlet.pull_chunk()
+        if timestamps:
+            print(timestamps, chunk)
 
 
 if __name__ == "__main__":
