+++ src/pylsl/examples/SendStringMarkers.py
